pyproject.toml
[project]
name = "end-to-end-ml-dl-project"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "apache-airflow-providers-databricks>=7.7.3",
    "apache-airflow[core]>=2.9.1",
    "cryptography>=45.0.7",
    "databricks-sql-connector>=4.1.2",
    "dbt-core>=1.10.10",
    "dbt-databricks>=1.10.9",
    "flask>=3.1.2",
    "flask-appbuilder>=5.0.1",
]
